requirements.txt
annotated-types==0.5.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==2.15.6
attrs==23.1.0
autopep8==2.0.4
black==23.9.1
Brotli==1.1.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.4.0
cryptography==41.0.4
cssselect2==0.7.0
dill==0.3.7
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-resized==1.0.2
django-weasyprint==2.2.1
djangorestframework==3.14.0
docstring-to-markdown==0.12
flake8==6.0.0
fonttools==4.45.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
lsprotocol==2023.0.0a3
lxml==4.9.3
mccabe==0.7.0
model-bakery==1.18.0
model-mommy==2.0.0
mypy==1.5.1
mypy-extensions==1.0.0
mypy-ls==0.5.1
nodeenv==1.8.0
oscrypto==1.3.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
phonenumbers==8.13.20
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydocstyle==6.3.0
pydyf==0.8.0
pyflakes==3.0.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pylint==2.17.5
pyls-isort==0.2.2
pylsp-mypy==0.6.7
pypdf==3.16.4
pyphen==0.14.0
pypng==0.20220715.0
pyright==1.1.326
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.1
python-lsp-ruff==1.5.1
python-lsp-server==1.7.4
pytoolconfig==1.2.5
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
rope==1.9.0
ruff==0.0.287
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
ujson==5.8.0
uritools==4.0.2
urllib3==2.0.6
weasyprint==60.1
webencodings==0.5.1
whatthepatch==1.0.5
wrapt==1.15.0
xhtml2pdf==0.2.11
yapf==0.40.1
zipp==3.16.2
zopfli==0.2.3
